Document cpu/endian.h

+++ libyaul/scu/bus/cpu/cpu/endian.h
@@ -9,22 +9,46 @@
 
 __BEGIN_DECLS
 
+/// @defgroup CPU_ENDIAN_DEFINES
+/// @defgroup CPU_ENDIAN_INLINE_FUNCTIONS
+
+/// @addtogroup CPU_ENDIAN_DEFINES
+/// @{
+
+/// @brief Convert from host byte order to big-endian order.
 #define cpu_htobe16(x)  ((uint16_t)(x))
+/// @brief Convert from host byte order to little-endian order.
 #define cpu_htole16(x)  cpu_bswap16((x))
+/// @brief Convert from big-endian order to host byte order.
 #define cpu_be16toh(x)  ((uint16_t)(x))
+/// @brief Convert from little-endian order to host byte order.
 #define cpu_le16toh(x)  cpu_bswap16((x))
 
+/// @brief Convert from host byte order to big-endian order.
 #define cpu_htobe32(x)  ((uint32_t)(x))
+/// @brief Convert from host byte order to little-endian order.
 #define cpu_htole32(x)  cpu_bswap32((x))
+/// @brief Convert from big-endian order to host byte order.
 #define cpu_be32toh(x)  ((uint32_t)(x))
 
+/// @brief Convert from little-endian order to host byte order.
 #define cpu_le32toh(x)  cpu_bswap32((x))
 
-/* Via __builtin_bswap32() and __builtin_bswap16(), GCC emits a function
- * with no use of the swap instructions */
+/// @}
 
-static inline uint32_t __always_inline
-cpu_bswap16(uint32_t x)
+/// @addtogroup CPU_ENDIAN_INLINE_FUNCTIONS
+/// @{
+
+/// @brief Swap the upper and lower bytes of a 2-byte value.
+///
+/// @details The `__builtin_bswap16()` function that GCC emits does not use the
+/// `swap.b` instruction.
+///
+/// @param x Value to be swapped.
+///
+/// @returns Swapped value.
+static inline uint16_t __always_inline
+cpu_bswap16(uint16_t x)
 {
         register uint32_t out;
 
@@ -35,6 +59,14 @@ cpu_bswap16(uint32_t x)
         return out;
 }
 
+/// @brief Swap the upper and lower 2-bytes of a 4-byte value.
+///
+/// @details The `__builtin_bswap32()` function that GCC emits does not use the
+/// `swap.w` instruction.
+///
+/// @param x Value to be swapped.
+///
+/// @returns Swapped value.
 static inline uint32_t __always_inline
 cpu_bswap32(uint32_t x)
 {
@@ -48,11 +80,17 @@ cpu_bswap32(uint32_t x)
         __asm__ volatile ("swap.b %[in], %[ot0]\n"
                           "swap.w %[it0], %[ot1]\n"
                           "swap.b %[it1], %[out]\n"
-                : [out] "=&r" (out), [ot0] "=r" (t0), [ot1] "=r" (t1)
-                : [in] "0" "r" (x), [it0] "1" (t0), [it1] "2" (t1));
+                : [out] "=&r" (out),
+                  [ot0] "=r" (t0),
+                  [ot1] "=r" (t1)
+                : [in] "0" "r" (x),
+                  [it0] "1" (t0),
+                  [it1] "2" (t1));
 
         return out;
 }
+
+/// @}
 
 __END_DECLS
